DROID-3742 fixes
